SpringBoot API
SpringBoot API implemented.
Run the SprinBoot project with IntelliJ IDEA and open the project as Maven.

Before runnning the project you have to insert some of your informations:

SpringBoot:
application.properties -> set the user who is connected to the database (maybe postgres) and the password.

Android Studio project:
RetrofitService.java -> Change the content of the String url with yours ip address (you can find by typing ipconfig on the terminal)

There's still one problem, when you register a new user the system will generate a message "Registration failed". But this is a "problem". The registration finish well and after the operation there's a new user on the database. We must check what's the problem and fix it.